requirements.txt
streamlit
numpy
pillow
tensorflow==2.13.0